test: autocompletion
