+++ packages/pnpm/src/shortHands.ts
@@ -1,0 +1,39 @@
+// tslint:disable
+export default {
+  's': '--reporter=silent',
+  'd': '--loglevel=info',
+  'dd': '--loglevel=verbose',
+  'ddd': '--loglevel=silly',
+  'L': '--latest',
+  'r': '--recursive',
+  'silent': '--reporter=silent',
+  'verbose': '--loglevel=verbose',
+  'quiet': '--loglevel=warn',
+  'q': '--loglevel=warn',
+  'h': '--help',
+  'H': '--help',
+  '?': '--help',
+  'usage': '--help',
+  'v': '--version',
+  'f': '--force',
+  'local': '--no-global',
+  'l': '--long',
+  'p': '--parseable',
+  'porcelain': '--parseable',
+  'prod': '--production',
+  'development': '--dev',
+  'g': '--global',
+  'S': '--save',
+  'D': '--save-dev',
+  'P': '--save-prod',
+  'E': '--save-exact',
+  'O': '--save-optional',
+  'C': '--dir',
+  'shrinkwrap-only': '--lockfile-only',
+  'shared-workspace-shrinkwrap': '--shared-workspace-lockfile',
+  'frozen-shrinkwrap': '--frozen-lockfile',
+  'prefer-frozen-shrinkwrap': '--prefer-frozen-lockfile',
+  'W': '--ignore-workspace-root-check',
+  'i': '--interactive',
+}
+// tslint:enable
