+++ packages/pnpm/test/index.ts
@@ -1,5 +1,6 @@
 ///<reference path="../../../typings/index.d.ts" />
 import './cli'
+import './completion.test'
 import './help.spec'
 import './install'
 import './link'
